Compiler pass-based restructure
