+++ src/compiler.rs
@@ -1,10 +1,15 @@
-use std::{collections::HashMap, ops::Range, slice::Iter};
+use std::{collections::HashMap, slice::Iter};
 
-use enumset::EnumSet;
-mod optimizations;
+mod preprocess;
+use preprocess::{preprocess, PreprocessData};
+mod first_pass;
+use first_pass::first_pass;
+mod final_pass;
+use final_pass::final_pass;
 
 pub use crate::parser::{ast::*, *};
 use crate::{Error, ErrorKind};
+
 
 /// Primitive bytecode needed to check action preconditions and apply effects to a "state"(memory).
 #[derive(Debug, PartialEq)]
@@ -155,60 +160,15 @@ pub struct CompiledAction<'src> {
     pub effect: Vec<Instruction>,
 }
 
-#[derive(Debug)]
-struct CompilerData<'src> {
-    requirements: EnumSet<Requirement>,
-    /// Type inheritance tree. Everything should end up being an object.
-    type_tree: HashMap<&'src str, Name<'src>>,
-    /// Source code position mapping of where types are declared.
-    type_src_pos: HashMap<&'src str, Range<usize>>,
-    /// Problem object map to Domain type.
-    object_to_type_map: HashMap<&'src str, Name<'src>>,
-    /// Domain type map to vector of objects. Note: `"object"` type must list all
-    /// problem objects.
-    type_to_objects_map: HashMap<&'src str, Vec<Name<'src>>>,
-    /// Source code position mapping of where problem objects are declared.
-    object_src_pos: HashMap<&'src str, Range<usize>>,
-    /// Mapping a vector of `[predicate, arg1, arg2, .., argN]` to a memory bit offset.
-    pub predicate_memory_map: HashMap<Vec<&'src str>, usize>,
-}
-
 /// Compile and optimize parsed [`Domain`] and [`Problem`] structures into a compiled problem ready for using in search methods.
 /// Load them from a file using [`parse_domain`] and [`parse_problem`] functions.
 pub fn compile_problem<'src>(
     domain: &Domain<'src>,
     problem: &Problem<'src>,
 ) -> Result<CompiledProblem<'src>, Error<'src>> {
-    validate_problem(domain, problem)?;
-    let compiler = map_objects(domain, problem)?;
-    let actions = create_concrete_actions(&compiler, domain)?;
-    let init = compile_init(&compiler, &problem.init)?;
-    let mut goal = Vec::with_capacity(32);
-
-    compile_precondition(&compiler, &problem.goal, None, &mut goal)?;
-    Ok(CompiledProblem {
-        memory_size: compiler.predicate_memory_map.len(),
-        actions,
-        init,
-        goal,
-    })
-}
-
-fn validate_problem<'src>(
-    domain: &Domain<'src>,
-    problem: &Problem<'src>,
-) -> Result<(), Error<'src>> {
-    use ErrorKind::*;
-    if problem.domain.1 != domain.name.1 {
-        return Err(Error {
-            input: None,
-            kind: MissmatchedDomain(domain.name.1),
-            chain: None,
-            range: problem.domain.0.clone(),
-        });
-    }
-
-    Ok(())
+    let preprocess_data = preprocess(domain, problem)?;
+    let first_pass_data = first_pass(preprocess_data, domain, problem)?;
+    final_pass(&first_pass_data, domain, problem)
 }
 
 /// Given a list of types, use a type to object map and generate all possible
@@ -317,436 +277,7 @@ where
     Ok(result)
 }
 
-fn compile_init<'src>(
-    compiler: &CompilerData<'src>,
-    init: &[Init<'src>],
-) -> Result<Vec<Instruction>, Error<'src>> {
-    let mut instructions = Vec::with_capacity(32);
-    for exp in init {
-        match exp {
-            Init::AtomicFormula(formula) => {
-                instructions.push(Instruction::And(0));
-                compile_name_negative_formula(compiler, formula, &mut instructions)?;
-            }
-            Init::At(_, _) => todo!(),
-            Init::Equals(_, _) => todo!(),
-            Init::Is(_, _) => todo!(),
-        }
-    }
-    Ok(instructions)
-}
 
-fn compile_term_atomic_formula<'src>(
-    compiler: &CompilerData<'src>,
-    af: &AtomicFormula<'src, Term<'src>>,
-    args: Option<&[(&Name<'src>, &Name<'src>)]>,
-    instructions: &mut Vec<Instruction>,
-    is_effect: bool,
-) -> Result<(), Error<'src>> {
-    use ErrorKind::{ExpectedName, ExpectedVariable, UndeclaredVariable};
-    use Term::*;
-    match af {
-        AtomicFormula::Predicate(name, params) => {
-            let mut call_vec = Vec::with_capacity(params.len() + 1);
-            call_vec.push(name.1);
-            for param in params {
-                match param {
-                    Variable(var) => {
-                        if args.is_none() {
-                            return Err(Error {
-                                input: None,
-                                kind: ExpectedName,
-                                chain: None,
-                                range: param.range(),
-                            });
-                        }
-                        let args = args.unwrap();
-                        let mut is_found = false;
-                        for arg in args {
-                            if var.1 == arg.0 .1 {
-                                call_vec.push(arg.1 .1);
-                                is_found = true;
-                                break;
-                            }
-                        }
-                        if !is_found {
-                            return Err(Error {
-                                input: None,
-                                kind: UndeclaredVariable,
-                                chain: None,
-                                range: var.0.clone(),
-                            });
-                        }
-                    }
-                    Name(name) => {
-                        if args.is_some() {
-                            return Err(Error {
-                                input: None,
-                                kind: ExpectedVariable,
-                                chain: None,
-                                range: param.range(),
-                            });
-                        }
-                        call_vec.push(name.1);
-                    }
-                    Function(_) => todo!(),
-                }
-            }
-            if let Some(offset) = compiler.predicate_memory_map.get(&call_vec) {
-                instructions.push(if is_effect {
-                    Instruction::SetState(*offset)
-                } else {
-                    Instruction::ReadState(*offset)
-                })
-            } else {
-                panic!("All variables matched, and predicate exists, but can't find this call vec memory offset")
-            }
-            Ok(())
-        }
-        AtomicFormula::Equality(_, _) => todo!(),
-    }
-}
-
-fn compile_name_atomic_formula<'src>(
-    compiler: &CompilerData<'src>,
-    af: &AtomicFormula<'src, Name<'src>>,
-    instructions: &mut Vec<Instruction>,
-) -> Result<(), Error<'src>> {
-    match af {
-        AtomicFormula::Predicate(name, args) => {
-            let mut call_vec = vec![name.1];
-            call_vec.extend(args.iter().map(|a| a.1));
-            if let Some(offset) = compiler.predicate_memory_map.get(&call_vec) {
-                instructions.push(Instruction::SetState(*offset));
-                Ok(())
-            } else {
-                // Check if predicate is defined, check of object is defined.
-                todo!()
-            }
-        }
-        AtomicFormula::Equality(_, _) => todo!(),
-    }
-}
-
-fn compile_gd<'src>(
-    compiler: &CompilerData<'src>,
-    gd: &GD<'src>,
-    args: Option<&[(&Name<'src>, &Name<'src>)]>,
-    instructions: &mut Vec<Instruction>,
-) -> Result<(), Error<'src>> {
-    match gd {
-        GD::AtomicFormula(af) => {
-            compile_term_atomic_formula(compiler, af, args, instructions, false)
-        }
-        GD::And(vec) => {
-            for gd in vec {
-                compile_gd(compiler, gd, args, instructions)?
-            }
-            Ok(())
-        }
-        GD::Or(_) => todo!(),
-        GD::Not(gd) => {
-            compile_gd(compiler, gd, args, instructions)?;
-            instructions.push(Instruction::Not);
-            Ok(())
-        }
-        GD::Imply(_) => todo!(),
-        GD::Exists(_) => todo!(),
-        GD::Forall(_) => todo!(),
-        GD::LessThan(_, _) => todo!(),
-        GD::LessThanOrEqual(_, _) => todo!(),
-        GD::Equal(_, _) => todo!(),
-        GD::GreatherThanOrEqual(_, _) => todo!(),
-        GD::GreaterThan(_, _) => todo!(),
-    }
-}
-
-fn compile_precondition<'src>(
-    compiler: &CompilerData<'src>,
-    precondition: &PreconditionExpr<'src>,
-    args: Option<&[(&Name<'src>, &Name<'src>)]>,
-    instructions: &mut Vec<Instruction>,
-) -> Result<(), Error<'src>> {
-    match precondition {
-        PreconditionExpr::And(vec) => {
-            for preconditions in vec {
-                compile_precondition(compiler, preconditions, args, instructions)?
-            }
-            instructions.push(Instruction::And(vec.len()));
-            Ok(())
-        }
-        PreconditionExpr::Forall(_) => todo!(),
-        PreconditionExpr::Preference(_) => todo!(),
-        PreconditionExpr::GD(gd) => compile_gd(compiler, gd, args, instructions),
-    }
-}
-
-fn compile_term_negative_formula<'src>(
-    compiler: &CompilerData<'src>,
-    formula: &NegativeFormula<'src, Term<'src>>,
-    args: Option<&[(&Name<'src>, &Name<'src>)]>,
-    instructions: &mut Vec<Instruction>,
-    is_effect: bool,
-) -> Result<(), Error<'src>> {
-    match formula {
-        NegativeFormula::Direct(af) => {
-            compile_term_atomic_formula(compiler, af, args, instructions, is_effect)
-        }
-        NegativeFormula::Not(af) => {
-            if is_effect {
-                instructions.push(Instruction::Not);
-            }
-            compile_term_atomic_formula(compiler, af, args, instructions, is_effect)?;
-            if !is_effect {
-                instructions.push(Instruction::Not);
-            }
-            Ok(())
-        }
-    }
-}
-
-fn compile_name_negative_formula<'src>(
-    compiler: &CompilerData<'src>,
-    formula: &NegativeFormula<'src, Name<'src>>,
-    instructions: &mut Vec<Instruction>,
-) -> Result<(), Error<'src>> {
-    match formula {
-        NegativeFormula::Direct(af) => compile_name_atomic_formula(compiler, af, instructions),
-        NegativeFormula::Not(af) => {
-            instructions.push(Instruction::Not);
-            compile_name_atomic_formula(compiler, af, instructions)
-        }
-    }
-}
-
-fn compile_fexp<'src>(
-    _: &CompilerData<'src>,
-    fexp: &FluentExpression<'src>,
-    instructions: &mut Vec<Instruction>,
-) -> Result<(), Error<'src>> {
-    match fexp {
-        FluentExpression::Number(n) => instructions.push(Instruction::Push(*n)),
-        FluentExpression::Subtract(_) => todo!(),
-        FluentExpression::Negate(_) => todo!(),
-        FluentExpression::Divide(_) => todo!(),
-        FluentExpression::Add(_) => todo!(),
-        FluentExpression::Multiply(_) => todo!(),
-        FluentExpression::Function(_) => todo!(),
-    }
-    Ok(())
-}
-
-enum SupportedFunctionOp {
-    INC,
-}
-fn function_op<'src>(
-    compiler: &CompilerData<'src>,
-    function: &FunctionTerm<'src>,
-    fexp: &FluentExpression<'src>,
-    op: SupportedFunctionOp,
-    instructions: &mut Vec<Instruction>,
-) -> Result<(), Error<'src>> {
-    compile_fexp(compiler, fexp, instructions)?;
-    if function.terms.len() == 0
-        && function.name.1 == "total-cost"
-        && compiler.requirements.contains(Requirement::ActionCosts)
-    {
-        use SupportedFunctionOp::*;
-        instructions.push(Instruction::ReadFunction(0)); // todo! map functions
-        match op {
-            INC => instructions.push(Instruction::Add),
-        }
-        instructions.push(Instruction::SetFunction(0));
-        Ok(())
-    } else {
-        Err(Error {
-            input: None,
-            kind: ErrorKind::UnsetRequirement(Requirement::ActionCosts.into()),
-            chain: None,
-            range: function.range(),
-        })
-    }
-}
-
-fn compile_effect<'src>(
-    compiler: &CompilerData<'src>,
-    effect: &Effect<'src>,
-    args: Option<&[(&Name<'src>, &Name<'src>)]>,
-    instructions: &mut Vec<Instruction>,
-) -> Result<(), Error<'src>> {
-    match effect {
-        Effect::And(vec) => {
-            for effect in vec {
-                compile_effect(compiler, effect, args, instructions)?
-            }
-            Ok(())
-        }
-        Effect::Forall(_) => todo!(),
-        Effect::When(_) => todo!(),
-        Effect::NegativeFormula(f) => {
-            instructions.push(Instruction::And(0));
-            compile_term_negative_formula(compiler, f, args, instructions, true)
-        }
-        Effect::Assign(_, _) => todo!(),
-        Effect::AssignTerm(_, _) => todo!(),
-        Effect::AssignUndefined(_) => todo!(),
-        Effect::ScaleUp(_, _) => todo!(),
-        Effect::ScaleDown(_, _) => todo!(),
-        Effect::Increase(function, fexp) => function_op(
-            compiler,
-            function,
-            &fexp.1,
-            SupportedFunctionOp::INC,
-            instructions,
-        ),
-        Effect::Decrease(_, _) => todo!(),
-    }
-}
-
-fn compile_basic_action<'src>(
-    compiler: &CompilerData<'src>,
-    args: Option<&[(&Name<'src>, &Name<'src>)]>,
-    action: &BasicAction<'src>,
-) -> Result<CompiledAction<'src>, Error<'src>> {
-    let mut precondition = Vec::new();
-    action
-        .precondition
-        .as_ref()
-        .and_then(|p| Some(compile_precondition(compiler, p, args, &mut precondition)))
-        .unwrap_or(Ok(()))?;
-    let mut effect = Vec::new();
-    action
-        .effect
-        .as_ref()
-        .and_then(|e| Some(compile_effect(compiler, e, args, &mut effect)))
-        .unwrap_or(Ok(()))?;
-    let args: Vec<Name> = if let Some(args) = args {
-        args.iter().map(|(_, o)| (**o).clone()).collect()
-    } else {
-        Vec::new()
-    };
-    Ok(CompiledAction {
-        name: action.name.clone(),
-        args,
-        precondition,
-        effect,
-    })
-}
-
-/// Action parameters type to object map results in permutations
-/// of objects that match given type. When checking if action can
-/// be accomplished we need to choose over which actions to perform it.
-/// Having one action per concrete object tuple allows us to use the same
-/// heuristic to decide if actions over some objects are more likely to
-/// accomplish goal than other actions over some objects. Alternative is
-/// to have heuristic function advise both order of actions and order of
-/// objects.
-/// # Arguments
-/// * `compiler` - all pre-computed maps maintained by the compiler
-/// * `domain` - PDDL Domain that defines all the actions needed to create.
-fn create_concrete_actions<'src>(
-    compiler: &CompilerData<'src>,
-    domain: &Domain<'src>,
-) -> Result<Vec<CompiledAction<'src>>, Error<'src>> {
-    let mut all_actions =
-        Vec::with_capacity(compiler.predicate_memory_map.len() * domain.actions.len() / 5);
-    for action in &domain.actions {
-        if let Action::Basic(action @ BasicAction { parameters, .. }) = action {
-            // Create action for all type-object permutations.
-            let actions = for_all_type_object_permutations(
-                &compiler.type_to_objects_map,
-                parameters.as_slice(),
-                |args| compile_basic_action(compiler, Some(args), action),
-            )?;
-            all_actions.extend(actions)
-        } else {
-            todo!()
-        }
-    }
-    Ok(all_actions)
-}
-
-/// Iterate through proper domain and problem structs to build type tree,
-/// object type map, and mapping of predicate calls to memory offsets.
-/// # Arguments
-/// * `domain` - PDDL Domain
-/// * `problem` - PDDL Problem
-fn map_objects<'src>(
-    domain: &Domain<'src>,
-    problem: &Problem<'src>,
-) -> Result<CompilerData<'src>, Error<'src>> {
-    let requirements = domain.requirements | problem.requirements;
-    let mut type_tree = HashMap::new();
-    let mut type_src_pos = HashMap::new();
-    // Iterate over types and build the type tree
-    for List { items, kind } in &domain.types {
-        if let Type::Exact(kind) = kind {
-            for item in items {
-                type_src_pos.insert(item.1, item.0.clone());
-                type_tree.insert(item.1, kind.to_owned());
-            }
-        } else {
-            todo!()
-        }
-    }
-    let mut object_to_type_map = HashMap::new();
-    let mut type_to_objects_map = HashMap::new();
-    let mut object_src_pos = HashMap::new();
-    type_to_objects_map.insert("object", Vec::new());
-    // Iterate over objects and build a Object to Type and Type to Object maps
-    for List { items, kind } in &problem.objects {
-        match kind {
-            Type::Exact(kind) => {
-                for item in items {
-                    object_src_pos.insert(item.1, item.0.clone());
-                    object_to_type_map.insert(item.1, kind.to_owned());
-                    let mut type_name = kind.1;
-                    while let Some(parent_type) = type_tree.get(type_name) {
-                        if type_to_objects_map
-                            .get_mut(type_name)
-                            .and_then(|vec: &mut Vec<Name>| Some(vec.push(item.to_owned())))
-                            .is_none()
-                        {
-                            type_to_objects_map.insert(type_name, vec![item.to_owned()]);
-                        }
-                        type_name = parent_type.1;
-                    }
-                    type_to_objects_map
-                        .get_mut("object")
-                        .and_then(|vec: &mut Vec<Name>| Some(vec.extend(items.iter().cloned())));
-                }
-            }
-            Type::None => {
-                object_src_pos.extend(items.iter().map(|i| (i.1, i.0.clone())));
-                type_to_objects_map
-                    .get_mut("object")
-                    .and_then(|vec: &mut Vec<Name>| Some(vec.extend(items.iter().cloned())));
-            }
-            Type::Either(_) => todo!(),
-        }
-    }
-    let mut predicate_memory_map = HashMap::new();
-    //Iterate over predicates and its objects and build a memory map
-    for AtomicFSkeleton { name, variables } in &domain.predicates {
-        for_all_type_object_permutations(&type_to_objects_map, &variables, |args| {
-            let mut call_vec = vec![name.1];
-            call_vec.extend(args.iter().map(|i| i.1 .1));
-            if !predicate_memory_map.contains_key(&call_vec) {
-                predicate_memory_map.insert(call_vec, predicate_memory_map.len());
-            }
-            Ok(())
-        })?;
-    }
-    Ok(CompilerData {
-        requirements,
-        type_tree,
-        type_src_pos,
-        object_to_type_map,
-        type_to_objects_map,
-        object_src_pos,
-        predicate_memory_map,
-    })
-}
 
 #[cfg(test)]
 pub mod tests {
@@ -780,29 +311,7 @@ pub mod tests {
         (domain, problem)
     }
 
-    #[test]
-    fn test_mapping() {
-        let (domain, problem) = get_tiny_domain();
-        let compiler = map_objects(&domain, &problem).unwrap();
-        assert_eq!(
-            compiler.predicate_memory_map,
-            HashMap::<Vec<&str>, usize>::from([
-                (vec!["a", "a"], 0),
-                (vec!["a", "b"], 1),
-                (vec!["a", "c"], 2),
-                (vec!["b", "a", "a"], 3),
-                (vec!["b", "b", "a"], 4),
-                (vec!["b", "c", "a"], 5),
-                (vec!["b", "a", "b"], 6),
-                (vec!["b", "b", "b"], 7),
-                (vec!["b", "c", "b"], 8),
-                (vec!["b", "a", "c"], 9),
-                (vec!["b", "b", "c"], 10),
-                (vec!["b", "c", "c"], 11),
-            ])
-        )
-    }
-
+    
     #[test]
     fn test_action() {
         use Instruction::*;
