--- src/compiler/optimizations.rs
@@ -1,1 +1,0 @@
-
